continue bmp exporting

+++ libft/custom/ft_index.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_index(char *haystack, char needle)
+#include <sys/types.h>
+
+ssize_t	ft_index(char *haystack, char needle)
 {
-	int		i;
+	ssize_t	i;
 
 	i = 0;
 	while (haystack[i] != '\0' && haystack[i] != needle)
